Moved sync configuration settings to its ovn class/file Removed unecessary references to WebExtensionMechanism (for non-extensions)
