+++ src/app/config.ts
@@ -1,7 +1,6 @@
 import { IHttpClient } from "crunchyroll-lib/models/http/IHttpClient";
 import { ContainerConstructor } from "crunchyroll-lib/utils/container";
 import ccontainer from 'crunchyroll-lib/config';
-import { WebExtensionMechanism } from "./storage/mechanism/WebExtensionMechanism";
 import container from "../config/inversify.config";
 import { IStorage, IStorageSymbol } from "./storage/IStorage";
 
@@ -32,7 +31,7 @@ class GlobalSettingsImpl implements IGlobalSettings {
   private _syncRes: boolean = false;
 
   get syncResolution(): boolean {
-    return WebExtensionMechanism.sync && this._syncRes;
+    return SyncConfig.sync && this._syncRes;
   }
 
   set syncResolution(val: boolean) {
@@ -66,3 +65,43 @@ class GlobalSettingsImpl implements IGlobalSettings {
 
 let global = new GlobalSettingsImpl() as IGlobalSettings;
 export default global;
+
+export class SyncConfig {
+  private static _sync: boolean = false;
+  private static _untempsync: boolean = false;
+  private static _isActive: boolean = false;
+  private static _canModifyActive: boolean = true;
+
+  static set isExtension(val: boolean) {
+    if (!this._canModifyActive)
+      throw "Trying to modify SyncConfig isExtension after it has been disabled!";
+    this._isActive = val;
+  }
+
+  static disableIsExtensionWrite(): void {
+    this._canModifyActive = false;
+  }
+
+  static get isExtension(): boolean {
+    return this._isActive;
+  }
+
+  static get sync(): boolean {
+    return this.isExtension && this._sync;
+  }
+
+  static set sync(val: boolean) {
+    this._sync = val;
+    if (browser) // only set if we can (and if is extension)
+      browser.storage.sync.set({"sync": val}); // no need to wait for it
+  }
+
+  static tempSync(val: boolean|null = null) {
+    if (val === null) { // reset
+      this._sync = this._untempsync;
+    } else {
+      this._untempsync = this._sync;
+      this._sync = val!;
+    }
+  }
+}
